+++ src/main/java/Duke/Duke.java
@@ -1,0 +1,33 @@
+package Duke;
+import Duke.command.*;
+import Duke.task.*;
+public class Duke {
+    private Storage storage;
+    private TaskList tasks;
+    private Ui ui;
+
+    public Duke(String filePath) {
+        ui = new Ui();
+        storage = new Storage(filePath);
+        tasks = new TaskList(storage.load());
+    }
+
+    public void run() {
+        ui.showWelcome();
+        boolean isExit = false;
+        while (!isExit) {
+            try {
+                String fullCommand = ui.readCommand();
+                Command c = Parser.parse(fullCommand);
+                c.execute(tasks, ui, storage);
+                isExit = c.isExit();
+            } catch (Exception e) {
+                ui.print_message(e.getMessage());
+            }
+        }
+    }
+
+    public static void main(String[] args) {
+        new Duke("database.ser").run();
+    }
+}
