+++ src/main/java/Duke/Parser.java
@@ -1,0 +1,57 @@
+package Duke;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import Duke.command.*;
+import Duke.task.*;
+public class Parser{
+    public static LocalDate parseDate(String date){
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        LocalDate date1 = LocalDate.parse(date, formatter);
+        return date1;
+    }
+    public static Command parse(String msg){
+        if (msg.equals("bye")) {
+            return new ExitCommand();
+        } else if (msg.equals("list")) {
+            return new ListCommand();
+        } else if (msg.length() >= 4 && msg.substring(0, 4).equals("todo")) {
+            return new AddCommand(new Todos(msg.substring(4)));
+
+        } else if (msg.length() >= 8 && msg.substring(0, 8).equals("deadline")) {
+            String[] strarr = processDeadlineMsg(msg.substring(8));
+            return new AddCommand(new Deadline(strarr[0], parseDate(strarr[1])));
+        } else if (msg.length() >= 5 && msg.substring(0, 5).equals("event")) {
+            String[] strarr = processEventMsg(msg.substring(5));
+            return new AddCommand(new Event(strarr[0], parseDate(strarr[1]),parseDate(strarr[2])));
+//                return new AddCommand(new Event(msg.substring(5)));
+        } else if (msg.length() > 4 && msg.substring(0, 4).equals("mark")) {
+            int number = Integer.parseInt(msg.substring(5));
+            return new MarkCommand(number);
+        } else if (msg.length() > 6 && msg.substring(0, 6).equals("unmark")) {
+            int number = Integer.parseInt(msg.substring(7));
+            return new UnmarkCommand(number);
+        } else if (msg.length() > 6 && msg.substring(0, 6).equals("delete")) {
+            int number = Integer.parseInt(msg.substring(7));
+            return new DeleteCommand(number);
+        } else {
+            return new DontknowCommand();
+        }
+    }
+    private static String[] processEventMsg(String msg){
+        String[] arr = new String[3];
+        String[] strarr = msg.split("/from");
+        arr[0] = strarr[0].trim();
+        String[] strArr = strarr[1].split("/to");
+        arr[1] = strArr[1].trim();
+        arr[2]=strArr[0].trim();
+        return arr;
+    }
+    private static String[] processDeadlineMsg(String msg){
+        String[] arr = new String[2];
+        String[] strarr = msg.split("/by");
+        arr[0] = (strarr[0].trim());
+        arr[1] = strarr[1].substring(0).trim();
+        return arr;
+    }
+}
