+++ src/main/java/Duke/command/Command.java
@@ -1,0 +1,8 @@
+package Duke.command;
+import Duke.*;
+import Duke.task.*;
+public abstract class Command{
+    public abstract void execute(TaskList tsk, Ui ui, Storage storage);
+    public abstract boolean isExit();
+
+}
