+++ src/main/java/Duke/task/Todos.java
@@ -1,0 +1,17 @@
+package Duke.task;
+
+public class Todos extends task {
+    public Todos(String msg){
+        super(msg);
+    }
+
+    @Override
+    public String toString(){
+        if (access_state()){
+            return "[T][X] "+ access_message();
+        }
+        else{
+            return "[T][ ] "+ access_message();
+        }
+    }
+}
